fix: vite polyfill
The vite polyfill which is https://guybedford.com/es-module-preloading-integrity#modulepreload-polyfill
under the hood should not be loaded in SSR context
When doing react SSR it can create transient error I could not reproduce this with
vue.
Loading the polyfill only in app.js which live_vue example already does seems the way to go.