LiveVue initialization change - `createLiveVue` as a way to encapsulate component resolution and vue setup
Previously `initializeVueApp` was not working in SSR mode, since it was declared in app.js which couldn't be imported by server bundle. Now it's in a separate file as `createLiveVue().setup()` and can be imported by both client and server bundles.

Previous way of initialization still works, but is deprecated and will be removed in v1.0.