Fix: ensure slots containing non-ASCII characters display correctly
On the Elixir side of things, the contents of slots are encoded into UTF-8
and then into Base64. Prior to this fix, the client side code was then
decoding them with `atob`, which can only really handle ASCII, leading to
anything with non-ASCII char displaying as an unexpected string of characters.

This change should fix that, by adding an extra step to decode the UTF-8 into
JS strings.